+++ lib/chess_web/game_live.html.heex
@@ -1,3 +1,4 @@
 <.live_component
   module={ChessWeb.BoardLive}
-  id={@chess_game_topic_id} />
+  id={@chess_game_topic_id}
+  pieces={@pieces} />
